Minimal Reproducible Example
This sets up a minimal reproducible example to demonstrate the problem I have with using `{{< include
section/somefile.qmd >}}` in a Quarto manuscript.

The `earthquakes` object (list/vector) is created in a child document `sections/introduction.qmd` but is not available
when it comes to processing it in a second child document `sections/analysis.qmd`.

+ How are objects shared between different child documents?
+ Why are each child documents processed in isolation?
+ Should I save everything at each stage and then load in subsequent stages?